Moved files and folders around, added Raylib Linux
Relocated files and folders, so that the Linux directory more closely matches that of the Windows directory.

Additionally modified the source code for the Raylib version to allow it to compile on Linux, which it fails to in a way that I'd also expect to happen on Windows, but I can't test that.

At the very least, the code has now been fixed to compile and function on Linux.